+++ src/Mods/AstralArena/ScriptExtender/Lua/AstralArena/Shared/Constants.lua
@@ -1,6 +1,7 @@
 return {
-    ModuleUUID = "3b7d59a3-a846-4ecd-9c5d-a7bfe3f1b84f",
-    DisplayVersion = "0.2.1-alpha.1",
+    ModuleUUID = "29c48c80-8777-f7b5-6bb8-376c1c5d8db6",
+    DisplayVersion = "0.3.0-alpha.1",
+    ArenaLevel = "AA_Arena_Main",
     SchemaVersion = 1,
     EntrantCount = 8,
     MaxPartySize = 4,